requirements.txt
langchain
langchain_community
langchain_groq
langchain-core
fastapi
uvicorn
streamlit
chromadb
mlflow
pypdf
python-dotenv
sentence-transformers
groq
pydantic
typing-extensions
python-multipart
openai
google-generativeai
huggingface_hub
langchain_openai
langchain_google_genai
langchain_huggingface
langchain-chroma